src/main/java/com/github/renatocardosoalves/decoder/controllers/UserController.java
package com.github.renatocardosoalves.decoder.controllers;

import com.github.renatocardosoalves.decoder.models.UserModel;
import com.github.renatocardosoalves.decoder.services.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@RequiredArgsConstructor
@RestController
@CrossOrigin(origins = "*", maxAge = 3600)
@RequestMapping("/users")
public class UserController {

    private final UserService userService;

    @GetMapping
    public ResponseEntity<List<UserModel>> findAll() {
        return ResponseEntity.ok(this.userService.findAll());
    }

    @GetMapping("/{userId}")
    public ResponseEntity<UserModel> findById(@PathVariable UUID userId) {
        return this.userService.findById(userId)
                .map(ResponseEntity::ok)
                .orElseGet(() -> ResponseEntity.notFound().build());
    }

}
